Add routiens fill_diag_1rdm and fill_sings_1rdm, for when RDMExcitLevel=1
These are the same as Fill_Diag_RDM and Fill_Sings_RDM routines except
without the 2-RDM bits being added in, pretty much.

Also, pass the one_rdm arrays to the relevant filling wrapper routines,
to allow this to be tested. All of these wrappers will be redone
soon, so this is onyl temporarily for testing.

Similarly, change some code in the rdm_general module, firstly for
testing, and seconly passing the one_rdm%matrix array, instead of
the rdm type itself.