Change old RDM code to use new one_rdm_t data structure.
This means that the natural orbital routines can be changed to use
the new one_rdm_t data structures, while still allowing the old RDM
code to use it.

To do this, all the old RDM filling routines have been modified to
use this structure.

This is hopefully sensible, since the 1-RDMs are stored in exactly
the same way in both the new and old codes.

Possible bug still existing for natural orbital calculations - needs
more investigation.