Check when spawn list about to become full, so that multiple communications can be performed as needed
This is only done for the finalising routines, which are called as one-offs,
so we don't have to worry too much about doing multiple extra communications
per iteration.

Also, change the size of the RDM arrays created in initialisation, and add
an extra RDM object two_rdm_recv_2. This is annoying but is necessary
to perform multiple communications now, when outputting the spinned RDMs,
unless we want to overwrite the main RDM or merge the two called routines
into one, which I don't want. I think this is the best compromise, though
it does require a bit more memory.