Correct calc_1rdms_from_spinfree_2rdms, which should have spatial orbital labels, not spin
We should use spatial labels because this routine takes in a spin-free
2-RDM and outputs a spin-free 1-RDM. It doesn't work if taking in
spinned 2-RDMs.

Also, correct a bug introduced a few commits ago when I changed spatial
to spin labels - an r became an j instead of an k. Now it changed back
to an r since we are using spatial orbitals labels.