Make sure that spawned RDM list has at least 50*nProcessors slots available
When runnning big CASSCF calculations, we have relatively few orbitals, but
many electrons and walkers. Therefore, previously the RDM spawned list was
being made very small, even though there are many spawns. In fact, for the
test I was doing, there was less than one slot assigned per processor in the
spawning list, so it was crashing instantly.

This change makes sure that there are at least 50 slots available for each
processor within a given spawned list. This is somewhat arbitrary, but even
if we run on 100,000 cores (very optimistic!) this would make the spawning
list only have 5 million entries, and so be less than 100MB (assuming only
1 RDM is sampled).