Introduce new functions for accessing global_determinant_data array
For coding up transition density matrices, we will want to be able to
access either just standard RDMs, just transition RDMs, or the combined
averaged signs and starting iteration of both.

Therefore, this introduces new routines, in the same manner as the old
ones, to deal with these more general cases.

The tests seem to be unchanged after this. For now, all calls to set_av*
or get_av* or set_iter* or get_iter* have just been changed to '_standard'
calls, so transition RDMs aren't coded up currently and nothing has
actually changed in the running.