Add new option to print out 1-RDMs from read in 2-RDM popsfiles
Also alter read_2rdm_popsfile to perform multiple communications, to
allow for cases where we don't have enough space to do the communication
in one go.

Also, if the shift hasn't started to vary, instead of just not reading the
RDM popsfile, read it and then set it to 0, which allows us to still write
out the 1-RDMs in such cases.