A few corrections from last commit, see comments below
Set tReadRDMs to .false. if they are zeroed after reading. This
prevents some discrepencies between results without this option.

Also, change 0.0_dp to 0_int_rdm when zeroing elements array.

Also, clear the spawn hash table after reading in the 2-RDM,
just in case (which is actually not a correction to account for
the last commit, but from before.