Introduced a min_tau option to limit tau-search
Introduced a new option:
min-tau (float)
in the Calc block of the input to introduce a minimun time-step
value in the automated tau-search.
This limits the dependence of the absolute worst case excitation,
which may happen so seldom that it's not warranted to hard-set
the time-step to this worst case scenario.

Future improvements: maybe turn off the tau-search after the
minimum is reached, but this may be no wanted, since we still
want to change the pSingles, pDoubles etc. quantities
Think about how to disentangle those processes..