intr. a on-the-fly H_ij/pgen histogramming
through the keyword:

frequency-analysis [int] [float]

a on-the-fly frequency analysis of the H_ij/pgen ratio can be
activated.
this works by sorting the ratio into equidistant bins from 0 to
the highest ratio of H_ij/pgen

todo: make it work in parallel, need a global bin and bounds list
but have to thinks about how to do that most efficiently..

an idea: store a integer of the list size on each core, get the
maximum over mpi. then make all the lists on the processors this
size and refill the store values into it and then allocate a final
list and sum in all values from the different cores!