finished implementing the new automated tau-search
based on the integrated histograms of the H_ij/pgen frequency
ratios, the new automated time-step search integrates up to
a set threshold all the contributions from the different
kind of excitations (singles, doubles, paralle, antiparallel and
additional ones in the non-symmetric guga approach) and determines
the new time-step, as opposed to the worst case approach, which
was in use until now.

but there still has to be some fine-tuning to be done since it
approaches the maximum time-step due to death events leading to
spawns of anti-particles.
and a lot of blooms happen.. maybe too much

AND there seems to be a bug in the GUGA approach that the maximum
time-step due to death events is way too high, so the algorithm
gets unstable.. maybe there is something wrong with the
diagonal matrix elements, which would also explain the jumps in the
total energy in the guga approach, when switching reference, which
shouldn't happen actually ... look into that!