fixed a bug in the new tau-search for sym GUGA
did pDoubles = 1.0_dp - pDoubles lol

this was also the bug why it crashed in this case!
had nothing to do with the diagonal elements, like
i suggested in my last commit!
this jumps in total reference energy also happen in the
determinant case, when the reference det gets changed!
so dont worry!