Various changes to allow transition RDMs to be calculated in singlerun, and bug fixes to singlerun
Singlerun transition RDMs are a bit of a problem, because you only have
one simulation contributing to normal RDMs, but two contributing to
transition RDMs. So in case the averaging arrays need to be length
2 per RDM, and in some cases length 1. To deal with this, just make
it 2 in all cases. This will then involve some replication, but
makes the code simpler and removes many if-statements. Singlerun
normal RDMs are biased anyway, so are not critical.

This also fixes problems in the singlerun RDM code, which were not
known about by me before now. The problem was basically that RDM
contributions were not always set to 0 after adding them in, so
contributions can be added in multiple times. Because of this
change, singlerun RDM tests are now failing, and new benchmarks
need to be set.