changed semistoch and trialwavefunction for guga
changed the implementation of the setup of the trial-wavefunction
and the semistochastic space for the guga implementation to make
it way more efficient than before.

avoided the unnecessary loops over all the possible state combinations
to setup the hamiltonian matrix elements, which in the guga case
are anyway calculated by applying the hamiltonian to a state and
then binary searching if the other state is in this list.

now the hamiltonian is only once applied to a state H|nI>
and then in the loop over <nJ| we look if it is in the excitations
list.

BUT: TODO the semistochastic one has an error still!
walker number explodes