Add contributions involving HF det in both ways, when using HPHFs
When sampling transition RDMs, so that the two states involved (a and
b) in the RDM are different, then we have to consider adding in
contributions to the RDM in both

\psi_HF^a \psi_i^b <D_HF|...|D_i>

and also

\psi_HF^b \psi_i^a <D_HF|...|D_i>

orders.

For the existing HPHF tests (which are onyl for standard RDMs), benchmarks
will have to be reset, since it does include contributions which are
different due to stohcastic variation, but are the same on average.