Many changes to 1-RDM finalising, to make it more flexible and to work with transition RDMs
The order and way that the various processing functions are called has
been refactored and made less messy and fixed.

We now don't calculate the norms of the 1-RDMs whenever they are
constructed directly from 2-RDMs, since they are already normalised.
This prevents things going wrong with transition RDMs, since they
shouldn't be normalised in the same way as  standard RDMs.

Also, correct the normalisation of transition RDMs when using
1-RDMs directly.

The old RDM code still needs to use the same functions, so copy
them over the rdms_old directory.