Remove the option to add certain contributions to the RDM symmetrically
This messes things up badly for transition RDMs, which are not symmetric. We could
include the option for non-transition RDMs, but this is very messy because of the
way that all signs for all RDMs are added in one go. Also, adding these symmetric
contributions is actually probably very bad when using a spawned RDM list, because
you have to add more contributions to the spanwed RDM list, but don't get any
extra statistics - all observables will give exactly the same value, and everything
is symmetrised at the end anyway.

This will alter test suite results, but only for the Hermiticity errors, which
obviously increase a bit, but that's OK.

There also appears to be a change to a singlerun RDM test in the energy, which
I will now investigate.