When adding in RDM contributions from stochastic spawning, correct spawning from bra to ket
When spawning from the bra to the ket, we have to be careful to not
add in RDM contributions as we do when normally, which accounts for
spawning from the ket to the bra. This never used to be a problem
for standard RDMs because they are symmetric, so it worked out OK,
but for transition RDMs (and probably also for complex RDMs of all
types) we need to be more careful. Correct this case and add a lot
of commenting to hopefully explain the problem.