Output Hermiticity differences for tRDMs, and change name convention
Outputting Hermiticity differences for tRDMs is very useful for the
test suite, as a way of making sure that RDM elements are added to
the correct side of the tRDM, since getting this wrong would not
affect any observable estimates, such as dipole moments.

Also, update the test suite's extract script to pick out these
tRDM estimates.

Also, change the name convention for tRDMs, as the '->' is a bit
cumbersome and possibly confusing (with transition going in the
wrong direction).