fixed a bug in new projected energy calc
the current_stepvector information was not correctly set up for the
CSFs before using this information to calculate the projected
energy.

but there is still something not quite right with the guga projected
energy.. also for old cases, like the "normal" h8 system which was
correct already in the past... what did change?