first implementation of hubbard + guga finished
first preliminary implementation of the guga implementation
for the built-in hubbard model finished!

tests for the 2x2, and 3x3 cubice lattices look promising and
correct. although efficiency compared to the determinental
implementation is a bit underwhelming..
have to improve on that!

and have to implement the histogramming tau-search also!
otherwise the time-step gets too low