logging changes to print frq_histograms
and small changes to the hubbard implementation to
better suite the guga approach for a different
ordering of the single-particle orbitals. (not finished
continue work at home at laptop!)

and if i am at it already i could also finally introduce
twisted boundary conditions if i am deep in the hubbard
case already.
maybe also look at implementing a 3-band hubbard model or
introduce the possibility of next-nearest neighbor hopping.

the easiest additional implementation should be the
twisted boundary conditions for the square and tilted lattice!