implemented open BC conditions
for the real-space(!) hubbard model, open boundary conditions
are now implemented.
for cubic lattices it is also possible to specify mixed open
and periodic boundary conditions.
and i also found the bug in the tilted cubic  lattice for the
GUGA implementation i guess.
it was due to the orbital reordering, benefical for the
momentum space implementation, but not for the real-space
implementation, and there it also lead to incorrect TMATs is seems