fix of real-space hubbard and open BC
fixed the implementation of the real-space hubbard code in NECI
and implemented open boundary conditions.

CAUTION: for now it only works with the square lattice.
The tilted real-space implementation needs some tweaking still.
for now the TMAt in the tilted case is set up with only
open boundary conditions and is not periodic anymore. Will
be fixed tomorrow.

to run the real-space hubbard model use

nosymgen

as the excitation generator without any of

kpoints
lattice-excitgen

as keywords and the keyword

real

in the SYSTEM Hubbard block of the input.

for the open boundary conditions use the

open-bc []

keyword with optional additional inputs [x,y]

specifying mixed open and periodic boundary condition
depending on the lattice dimensions.

SIDE-NOTE:
this is only implemented for the 1- and 2- dimensional
hubbard model

to use open BC use the keyword