changed some default behaviors
changed the default behavior of t_guga_mateles to be .true.
except when turned off by the keyword

no-guga-mateles

in the CALC block of the input.

changed the default behavior of t_read_probs to be .true.
except when turned off by the keyword

no-read-probabilities

in the CALC block of the input.

kept the old keyword

guga-mateles
read-probabilities

for compatibility with old input files for now.
but they will be removed in the future!