removed the orbital noreorder and added tests!
this orbital reordering in the molecular calculations is not
so easy to implement as i expected! since i would have to
change how the matrx elements get accessed.
a better was is to use the local ordering in the case of
a local basis or the hubbard model or use a strict
energy ordering for SCF bases! but this should optimally already
be done at the FCIDUMP level!

and i added some testsuite tests specifically for GUGA, but which
have to be benchmarked still! but do this at the office machine!