Added new stats file EXLEVELStats
Added Boolean input keyword EXLEVEL to the LOGGING block.  If present,
the L0, L1, and L2 norms of the sum of walker weights for each
possible excitation level are logged to file "EXLEVELStats".  This is
useful for visualizing the evolution of the wave function with
projection time and investigating changes in walker dynamics arising
from modifications to the sampling algorithm.

NB, the different L-norms quantify the collective weight of each
excitation level in different ways:
* the L0 norm is the number of occupied determinants,
* the L1 norm is the number of walkers, and
* the L2 norm is the square root of the sum of squared CI
  coefficients, which can be interpreted as a "collective CI
  coefficient" of the excitation level.