intermediate bug commit:
there is a bug with using RDMs and k-point symmetry, due to the
storing of the k-point symmetries as powers of 2!

this crashes the accessing of the symmetry labels in the rdm
diagonalization!

in the GUGA branch i fixed that! so i have to apply the same
changes here to be able to use RDMs with k-point symmetry!