moved some definitions into tau_search_hist module
and the main commit message which i left out yesterday:

the new histogramming tau-search option does not determine the
time-step from the worst case H_ij / pgen ratio as it is done
right now, but histograms all the H_ij / ratios for the
different type of excitations(singles, doubles, parallel etc.)
and integrates this histogram up to a specified percentage and
determines the time-step from this upper bound ratio and also
adjusts the pSingles, pParallel quantities

to activate this option use:
new-tau-search [cutoff] [n_bins] [max_ratio]
or
hist-tau-search [cutoff] [n_bins] [max_ratio]

in the Calc block of the input.
 cutoff		float, which determines the percentage the histograms
 		are integrated up to. default: 0.999

 n_bins		the number of bins used for histogramming. Be careful
 		to not use to many bins, which may cause problems in
		the MPI communication. default: 100000

 max_ratio	the maximum (H_ij / ratio) which is histogrammed.
 		ratios exceeding this quantitiy are counted and
		a warning will be printed if this exceeds a
		threshold (to be introduced in future commits!)
		default: 10000.0

for now, these histograms are printed out by default at the end of
a calculation! in the file frequency_histogram_*
this might be changed in future commits!

the histogramming tau-search, for now, is stopped until one of the
histograms is "full", meaning that in the integration an integer
overflow occurs! then the final obtained time-step and pSingles etc.
are used for the rest of the calculation.

I also changed the default behavior for restarted runs, when
reading in the time-step: until now only the time-step was used, but
if one chose to continue adapting it in the standard tau-search,
the pSingles, etc. from the previous calculation were NOT used.
which can cause quite som instability in the restarted calcs!
especially for localized orbitals, where pSingles is much higher
than in localised orbitals!
this is now changed, that also pSingles etc. are used from a
previous calculation.

in combination with this histogramming tau-search a second option
is introduced:
because now the time-step is not limited by the worst case
excitation, it can happen that more bloom happen again, for
excitation with a worse H_ij/pgen ratio: so the keyword:

truncate-spawns [n_trunc]

in the CALC block reduces this spawns to the specified [n_trunc]
value. default: 3.0

BUT: i have to admit this is some sort of uncontrolled approximation,
but from experience, this does not seem to influence the results
in a noticable way!

from my tests this new tau-search can improve the obtained time-step
by up to an order of magnitude!

BTW: if the hist-tau-search and the standard tau-search are tried to
be used, the standard one is TURNED OFF in the initialisation!