added an option to mix the ratios!
instead of taking only the new quantity instead mix the
old and the new ones for time-step, psingles etc.
new = (1-x)*old + x*new

but i just realised from the first step, that there is
something going wrong, so i have to investigate that.

AND: i also increased the threshold of how many excitations
i need to histogram, until we start to vary the quantities
from 50 to 5000.

But i could be that this causes the problems too.. check that!