changed pgen lower bound and introduced
matrix upper bound.

in the part-exact version of the excitation generators some
strange lower bound for the pgens was set, which i removed now.

and another thing is: integrals below umateps are screened in the
calculation, but this bound is not applied to parallel excitations
I tried doing that now, but probably should talk with Ali about this

and this could still not solve the problem maybe! since it could
still be that the pgens are just too high and this whole approach
ist jost not as efficient as thought.
and remember: we are using only a 0.1 grid! this is not so bad
actually! from the perspective of the ratios!
i could use a finer grid! to better differentiate the actual
H_ij/pgen ratio!