fixed some issues
first of all fixed the NaN pgen issue without the
lower limits in the 4ind-weighted-2 excit gen.

then to compare different implementations:
introduced a temporal new excitation generator
keyword:
 4ind-weighted-unbound

to indicate the unbound implementation of the
4ind-weighted-2 excit gen.

then i also introduced a new limit for matrix elements
similar to UMATEPS when ignoring 2-body integrals when the
FCIDUMP is read in.
Because in my opinion this should be handled on an equal footing.

in this sense all the matrix elements, singles and (anti-) parallel
excitations are truncated in the same way.
and also pgens below this threshold are set to zero to improve
efficiency.

to test the effect of this i introduced a new keyword in the CALC
block:
 matrix-cutoff [float]

 with an optional float input as the cutoff value.
by default if no specific input is given it is the UMATEPS value
of 1.0e-8