rebenchmarked test_suite
with ifort/15.0.3 and mpi.intel/5.0.3
config on altest.

i cross tested against gfortran on altest and on my laptop
and on ifort configurations on allogin1 and on the fkf-office pc.

now the test_suite should be (hopefully) working again!

EXCEPT:
there is a problem with the hdf5 usage on altest, so i moved
the hdf5 tests to the test_suite/hdf5/ folder which currently is
NOT TESTED by default.

I will try to fix this as soon as possible.

But this is only a compilation issue, since on allogin, where the
hdf5 compilation works those tests, with correct benchmarks work!