now it is possible to switch between
the old and the less mpi heavy tau search with the intermediate
keyword:
test-hist-tau

in the calc block.

i have to test if the results are consistent with the old
implementation for bigger calcs!

PS: I think, since i know the node-specific total sums of the
histograms i can do the node-specific integration in such a way that
i get the exact same results as in the old implementation, but
i would need to communicate the sums additionally..
so i am not sure if this is not counterproductive..