first commit of alis back-spawn idea:
in this method we try to force the non-initiators to more
likely spawn to already occupied determinants.

the first rudimentary implementation of this idea tries to do
that by only picking electrons, which are in the virtual orbitals
of the reference determinant for non-initiator dets.

this will more likely reduce the excitation level of the determinant
or worst-case leave it at the same excitation level and thus increase
the chance to spawn to an already occupied determinant

this method is activated by the

back-spawn

keyword in the CALC block of the input and ofc only works if the
initiator methods is also turned on.

for now this is only implemented for excitation generators of
type 4ind-weighted-2, because i am too lazy to change stuff all
over the place.
i still need to do testing if it actually works though