fixed a few bug in the back-spawn implementation
and also tried to implement it for the hubbard models.

but i have to fix some circular dependencies first and also
optimize the hubbard excitation generation i guess

while doing that i noticed that the real-space hubbard excitation
generation is probably implemented in a really inefficient way.
and i should probably optimize that too!