added the option to pick the first orbital (a)
from the occupied orbitals in the reference determinant uniformly.
this has 2 reasons:
first alis idea was, that that may increase the matrix elements
and second it increases the chance to de-excite and thus increase the
chance to hit already occupied determinants.

this feature is enabled by the keyword:

back-spawn-occ-virt [integer]

in the CALC block, again with an integer, to start the feature later in
the simulation.