added an optimal time-step determination for
the real- and k-space hubbard model.

BUT: this is pretty high and has to be used cautious!
it is activated when using the histogramming tau-search.

but it is quite big and makes the calculation a bit unstable.
but i do not know why exactly.

AND: the real-space hubbard is implemented really crappy i realized!
i have to optimize that in the future