added new combination of the flex and occ-virt
back-spawning method. which enables us to decrease the
excitation level of a non-initiator definetly.

it is activated by an additional possible integer

back-spawn-flex [delay] [level]

which can be 1 or 2.

for 1 the excitation level usually is decreased by 1
and for 2 it can be even ensured to be decreased by 2

but this depends which electrons are picked. if they are in the
occupied or virtual manifold of the reference determinant

some testing still has to be done!