fixed a small bug in the setup for the
back-spawn method and i figured out why there is
a discprepancy in some of he old test data for back-spawn
and the newly excitation generator version:
it is because i actually fixed a bug, where the old implementation
was not using symmetries correctly for single excitations!
so I have to rerun those tests!