implemented the k-space hubbard + back-spawn in
a seperate exciation generator similar to the
ueg one. also added unit tests and fixed some small
issues in the rest of the code.

i realized though that there are even 3 different UEG
excitation generator implementations..
and i missed the newest one.. and when HPHF + UEG is
used in the new implementation it probably does not produce the
expected output.