small intermediate commit to fix failing
testsuite on altest in the debug + back-spawn + hphf
implementation.

BUT BE CAREFUL! this fix is only intermediate, since the
actual bug is not fixed!
actually the back-spawn + hphf is broken in master!
BUT this is fixed in the back-spawn branch, which i will soon
merge back into master, after the testsuite finishes for it!