add cepa utility to the rest of NECI
activate it by the keyword in the CALC block of the input:

cepa [cepa_method]

with cepa method being:
0
aqcc
acpf

0 is the default and cepa(1) and cepa(3) still have to be implemented.

this is the most basic implementation for now which just adjusts the
amplitudes of the singles and doubles, but does not take into account
any occupation measure of higher excitations and also does not change
any of the amplitudes of higher excitations directly.

and remember the cepa-shift formulas are derived for closed-shell
reference systems!