major changes to the cc-amplitudes implementation
changed double storage back to the full storage of all possible
doubles, since there could be more double than sampled in the
wavefunction, just from the singles amplitudes and for the
quadruple calculation i also need all the possible doubles
efficiently.

also fixed the quadrupels amplitude calculation, since i have to
account for all the permutations leading to the same quadrupel
excitation.
because of that quadrupels are now stored in a hash-table fashion.

but for now, only the T2^2 influence on the quadrupels is
calculated. the singles and triples is left out.
but with the hash-storage this could be implementetd now too..

another way would be to correct the estimated ammount of
quadrupels just by dividing by 18, which is the number of
possible permutations.. lets see