fix for HPHF + Transcorrelated
there was and is a major issue with hermiticity
in the current implementation. For the transcorrelated approach,
where the hamiltonian is not hermitian essentially almost all
matrix element calculation calls are in the wrong order!!

this is a hacky approach to fix this issue in the real-space
hubbard implementation, but this must be done more thorouhly
and must be generalised for other systems.

and especially the whole HPHF implementation needs a rework and
clean up!