fix hubbard transcorr + hphf issues..
the correct matrix elements atleast. I still have to assure that
the pgens and matrix elements are now correctly related in the case
of transcorrelation, or if i do it exactly the wrong way around(I
think i do that). this would mean i have to change the way the
projected energy is calculated and undo all the changes in the other
parts of the code.. to address hermiticity