start of a new k-space hubbard implementation
in the long run i want to rework the whole k-space hubbard
implementation in this branch.
But for the beginning the most important part is, that we want
to add the 1- and 2-body transcorrelated hamiltonians here.
for that we need to change the matrix element calculation and also
in the excitation generation use cumulative sums to weight towards
bigger matrix elements, with the transcorrelation factor taken into
account.
And i guess it is better to do this nice and clean in one place,
instead of messing aroung in all the old code for the momentum
space hubbard model..