further work on the hopping transcorrelation
implementation for the real-space hubbard model.
implemented the matrix elements and the umat setup now.

runtime testing for bugs still need to be done!
and i have to work on the lattice implementation to
additionally store the real-space coordinates/vectors
also, to do the "fourier transform" of the dispersion
relation correctly