first workings in analysing the point group
symmetry of the ground state wavefunction(when no point groups
are used) bases on applying symmetry operations on the wavefunction
and analysing the phase between the original and obtainend
wavefunction.

it is activated by the
symmetry-analysis [rotation/mirror/inversion] [degree/axis]
symmetry-states [read/pop] [n_states]

keywords in the Logging Block of the input.

the main functionality is in the analyse_wf_symmetry.F90
file and unit tests still need to be written.

IMPORTANT: I need to check if the fix of Khaldoon to the
parity calculation in the templated quicksort algorithm is in
any of his branches or master already. If not, I have to still
fix that for my implementation, since we rely on it!