Make it so that death can be performed *after* communication and merging of spawns
This is important if we want to use spawns to calculate various
replica-based estimators without using RDMS. For example, the
variational energy and variance of the energy can be calculated
by (essentially) taking dot products of the *merged* spawning
array and the walker array *before* death. So this ordering is
more appropriate for this situation.

Note that I give an option to turn this change off. For integer
coefficients, the death step uses random numbers, so this change
breaks the tests. The tests could be reset but some are currently
failing, so for now I just turn this option off. Similarly, it
seems to break the back spawn option. I don't know how this works
so just turn it off in that case for now.