Routine to set initiator flags based on whether the spawned-to site is occupied already
For a time step of 1, and using preconditioning, all walkers are
killed entirely each iteration, which can maess with initiator
if the flags are not set early enough. This was previously done when
finding replica estimates for efficiency. But now this option can be
turned off, so this is added for care.