Add an option to perform walker death before communication, from the input file
This is used to correct two tests:
mneci/rdm/N_open_shell
mneci/rdm/BeH_open_shell

Both of these tests have instances where the time step is so large that
anti-particles are created. This causes the walker to be removed
from the walker list and hash table in the death step. If this is
done during the loop over all walkers, then this causes walkers
to be places in different places in the list, which changes the Markov
chain. This isn't a bug, but breaks these two tests. Allowing this
option and adding it to the test inputs corrects this.

Note that these two tests are still failing, but this is due to
a problem on master which should be fixed when it is merged.